+++ docs/source/conf.py
@@ -17,12 +17,30 @@ import sys
 import os
 import json
 import pprint
+import subprocess as sp
 
 script_dir = os.path.dirname(os.path.realpath(__file__))
 
 ################################################################################
+
+cmd = "git branch | grep '*' | cut -f 2 -d \ " \
+    + " | sed -e 's/readthedocs.tutorial/tutorial/g' " \
+    + " | sed -e 's/readthedocs/release/g'"
+mytag = sp.Popen(cmd, shell=True, stdout=sp.PIPE).stdout.read().strip()
+
+if 'detached' in mytag :
+    cmd = "git branch | grep '*' | cut -f 2 -d '/' | cut -f 1 -d ')'" \
+        + " | sed -e 's/readthedocs.tutorial/tutorial/g' " \
+        + " | sed -e 's/readthedocs/release/g'"
+    mytag = sp.Popen(cmd, shell=True, stdout=sp.PIPE).stdout.read().strip()
+
+tags.add (mytag)
+
+################################################################################
 ##
 print "* Generating resource configuration docs: resources.rst"
+print "* using tag: %s" % mytag
+print "* Generating code example list: examples.rst"
 
 try:
     os.remove("{0}/resources.rst".format(script_dir))
@@ -36,6 +54,9 @@ with open("{0}/resources.rst".format(script_dir), "w") as resources_rst:
 
         if example.endswith(".json") is False:
             continue # skip all non-python files
+
+        if example.startswith("aliases") is True:
+            continue # skip alias files
 
         print " * %s" % example
 
@@ -53,14 +74,21 @@ with open("{0}/resources.rst".format(script_dir), "w") as resources_rst:
                     default_queue = resource_config["default_queue"]
                 except Exception, ex:
                     default_queue = None
+
                 try:
                     working_dir = resource_config["default_remote_workdir"]
                 except Exception, ex:
                     working_dir = "$HOME"
+
                 try:
                     python_interpreter = resource_config["python_interpreter"]
                 except Exception, ex:
                     python_interpreter = None
+
+                try:
+                    access_schemas = resource_config["schemas"]
+                except Exception, ex:
+                    access_schemas = ['n/a']
 
                 resources_rst.write("{0}\n".format(resource_key))
                 resources_rst.write("{0}\n\n".format("-"*len(resource_key)))
@@ -73,8 +101,9 @@ with open("{0}/resources.rst".format(script_dir), "w") as resources_rst:
                 resources_rst.write("================== ============================\n")
                 resources_rst.write("``queue``               {0}\n".format(default_queue))
                 resources_rst.write("``sandbox``             {0}\n".format(working_dir))
+                resources_rst.write("``access_schema``       {0}\n".format(access_schemas[0]))
                 resources_rst.write("================== ============================\n\n")
-
+                resources_rst.write("Available schemas: ``{0}``\n\n".format(', '.join(access_schemas)))
 
                 resources_rst.write(":download:`Raw Configuration file: {0} <../../src/radical/pilot/configs/{0}>`\n\n".format(example))
 ##