+++ examples/getting_started_local.py
@@ -61,125 +61,99 @@ def wait_queue_size_cb(umgr, wait_queue_size):
 #
 if __name__ == "__main__":
 
-    try :
+    # prepare some input files for the compute units
+    os.system ('hostname > file1.dat')
+    os.system ('date     > file2.dat')
 
-        # prepare some input files for the compute units
-        os.system ('hostname > file1.dat')
-        os.system ('date     > file2.dat')
+    # Create a new session. A session is the 'root' object for all other
+    # RADICAL-Pilot objects. It encapsulates the MongoDB connection(s) as
+    # well as security credentials.
+    session = rp.Session()
+    print "session id: %s" % session.uid
 
-        # Create a new session. A session is the 'root' object for all other
-        # RADICAL-Pilot objects. It encapsulates the MongoDB connection(s) as
-        # well as security credentials.
-        session = rp.Session()
-        print "session id: %s" % session.uid
+    # Add a Pilot Manager. Pilot managers manage one or more ComputePilots.
+    pmgr = rp.PilotManager(session=session)
 
-        # Add a Pilot Manager. Pilot managers manage one or more ComputePilots.
-        pmgr = rp.PilotManager(session=session)
+    # Register our callback with the PilotManager. This callback will get
+    # called every time any of the pilots managed by the PilotManager
+    # change their state.
+    pmgr.register_callback(pilot_state_cb)
 
-        # Register our callback with the PilotManager. This callback will get
-        # called every time any of the pilots managed by the PilotManager
-        # change their state.
-        pmgr.register_callback(pilot_state_cb)
+    # Define a 4-core local pilot that runs for 10 minutes and cleans up
+    # after itself.
+    pdesc = rp.ComputePilotDescription()
+    pdesc.resource = "local.localhost"
+    pdesc.runtime  = 5 # minutes
+    pdesc.cores    = 1
+    pdesc.cleanup  = True
+    pdesc.sandbox  = "/tmp/tmp_sandbox/test/test/"
 
-        # Define a 4-core local pilot that runs for 10 minutes and cleans up
-        # after itself.
-        pdesc = rp.ComputePilotDescription()
-        pdesc.resource = "local.localhost"
-        pdesc.runtime  = 5 # minutes
-        pdesc.cores    = 1
-        pdesc.cleanup  = True
+    # Launch the pilot.
+    pilot = pmgr.submit_pilots(pdesc)
 
-        # Launch the pilot.
-        pilot = pmgr.submit_pilots(pdesc)
+    # Combine the ComputePilot, the ComputeUnits and a scheduler via
+    # a UnitManager object.
+    umgr = rp.UnitManager(
+        session=session,
+        scheduler=rp.SCHED_DIRECT_SUBMISSION)
 
-        # Combine the ComputePilot, the ComputeUnits and a scheduler via
-        # a UnitManager object.
-        umgr = rp.UnitManager(
-            session=session,
-            scheduler=rp.SCHED_DIRECT_SUBMISSION)
+    # Register our callback with the UnitManager. This callback will get
+    # called every time any of the units managed by the UnitManager
+    # change their state.
+    umgr.register_callback(unit_state_cb, rp.UNIT_STATE)
 
-        # Register our callback with the UnitManager. This callback will get
-        # called every time any of the units managed by the UnitManager
-        # change their state.
-        umgr.register_callback(unit_state_cb, rp.UNIT_STATE)
+    # Add the previously created ComputePilot to the UnitManager.
+    umgr.add_pilots(pilot)
 
-        # Add the previously created ComputePilot to the UnitManager.
-        umgr.add_pilots(pilot)
+    # Create a workload of ComputeUnits (tasks). Each compute unit
+    # uses /bin/cat to concatenate two input files, file1.dat and
+    # file2.dat. The output is written to STDOUT. cu.environment is
+    # used to demonstrate how to set environment variables within a
+    # ComputeUnit - it's not strictly necessary for this example. As
+    # a shell script, the ComputeUnits would look something like this:
+    #
+    #    export INPUT1=file1.dat
+    #    export INPUT2=file2.dat
+    #    /bin/cat $INPUT1 $INPUT2
+    #
+    cuds = []
+    for unit_count in range(0, 16):
+        cud = rp.ComputeUnitDescription()
+        cud.executable    = "/bin/sh"
+        cud.environment   = {'INPUT1': 'file1.dat', 'INPUT2': 'file2.dat'}
+        cud.arguments     = ["-l", "-c", "cat $INPUT1 $INPUT2"]
+        cud.cores         = 1
+        cud.input_staging = ['file1.dat', 'file2.dat']
 
-        # Create a workload of ComputeUnits (tasks). Each compute unit
-        # uses /bin/cat to concatenate two input files, file1.dat and
-        # file2.dat. The output is written to STDOUT. cu.environment is
-        # used to demonstrate how to set environment variables within a
-        # ComputeUnit - it's not strictly necessary for this example. As
-        # a shell script, the ComputeUnits would look something like this:
-        #
-        #    export INPUT1=file1.dat
-        #    export INPUT2=file2.dat
-        #    /bin/cat $INPUT1 $INPUT2
-        #
-        cuds = []
-        for unit_count in range(0, 16):
-            cud = rp.ComputeUnitDescription()
-            cud.executable    = "/bin/sh"
-            cud.environment   = {'INPUT1': 'file1.dat', 'INPUT2': 'file2.dat'}
-            cud.arguments     = ["-l", "-c", "cat $INPUT1 $INPUT2"]
-            cud.cores         = 1
-            cud.input_staging = ['file1.dat', 'file2.dat']
+        cuds.append(cud)
 
-            cuds.append(cud)
+    # Submit the previously created ComputeUnit descriptions to the
+    # PilotManager. This will trigger the selected scheduler to start
+    # assigning ComputeUnits to the ComputePilots.
+    units = umgr.submit_units(cuds)
 
-        # Submit the previously created ComputeUnit descriptions to the
-        # PilotManager. This will trigger the selected scheduler to start
-        # assigning ComputeUnits to the ComputePilots.
-        units = umgr.submit_units(cuds)
+    # Register also a callback which tells us when all units have been
+    # assigned to pilots
+    umgr.register_callback(wait_queue_size_cb,   rp.WAIT_QUEUE_SIZE)
 
-        # Register also a callback which tells us when all units have been
-        # assigned to pilots
-        umgr.register_callback(wait_queue_size_cb,   rp.WAIT_QUEUE_SIZE)
+    # Wait for all compute units to reach a terminal state (DONE or FAILED).
+    umgr.wait_units()
 
-        # Wait for all compute units to reach a terminal state (DONE or FAILED).
-        umgr.wait_units()
+    print 'units all done'
+    print '----------------------------------------------------------------------'
 
-        print 'units all done'
-        print '----------------------------------------------------------------------'
+    for unit in units :
+        unit.wait ()
 
-        for unit in units :
-            unit.wait ()
+    for unit in units:
+        print "* Task %s (executed @ %s) state %s, exit code: %s, started: %s, finished: %s, stdout: %s" \
+            % (unit.uid, unit.execution_locations, unit.state, unit.exit_code, unit.start_time, unit.stop_time, unit.stdout)
 
-        for unit in units:
-            print "* Task %s (executed @ %s) state %s, exit code: %s, started: %s, finished: %s, stdout: %s" \
-                % (unit.uid, unit.execution_locations, unit.state, unit.exit_code, unit.start_time, unit.stop_time, unit.stdout)
+    # Close automatically cancels the pilot(s).
+    session.close ()
 
-        # Close automatically cancels the pilot(s).
-        session.close ()
-
-        # delete the test data files
-        os.system ('rm file1.dat')
-        os.system ('rm file2.dat')
-
-    except Exception as e :
-        # Something unexpected happened in the pilot code above
-        print "caught Exception: %s" % e
-
-    except (KeyboardInterrupt, SystemExit) as e :
-        # the callback called sys.exit(), and we can here catch the
-        # corresponding KeyboardInterrupt exception for shutdown.  We also catch
-        # SystemExit (which gets raised if the main threads exits for some other
-        # reason).
-        print "need to exit now: %s" % e
-
-    finally :
-        # always clean up the session, no matter if we caught an exception or
-        # not.
-        print "closing session"
-        session.close ()
-
-        # the above is equivalent to
-        #
-        #   session.close (cleanup=True, terminate=True)
-        #
-        # it will thus both clean out the session's database record, and kill
-        # all remaining pilots (none in our example).
-
+    # delete the test data files
+    os.system ('rm file1.dat')
+    os.system ('rm file2.dat')