+++ src/radical/pilot/compute_unit_description.py
@@ -156,8 +156,9 @@ class ComputeUnitDescription(attributes.Attributes) :
       # self._attributes_register(START_AFTER,      None, attributes.STRING, attributes.VECTOR, attributes.WRITEABLE)
       # self._attributes_register(CONCURRENT_WITH,  None, attributes.STRING, attributes.VECTOR, attributes.WRITEABLE)
 
-        self._attributes_register_deprecated ('input_data',  'input_staging',  flow=self._DOWN)
-        self._attributes_register_deprecated ('output_data', 'output_staging', flow=self._DOWN)
+      # disabled deprecated attributes
+      # self._attributes_register_deprecated ('input_data',  'input_staging',  flow=self._DOWN)
+      # self._attributes_register_deprecated ('output_data', 'output_staging', flow=self._DOWN)
 
         # explicitly set attrib defaults so they get listed and included via as_dict()
         self.set_attribute (KERNEL,         None)