+++ src/radical/pilot/controller/output_file_transfer_worker.py
@@ -107,8 +107,6 @@ class OutputFileTransferWorker(threading.Thread):
                     # AM: The code below seems wrong when BULK_LIMIT != 1 -- the
                     # compute_unit will be a list then I assume.
                     try:
-                        log_messages = []
-
                         # We have found a new CU. Now we can process the transfer
                         # directive(s) wit SAGA.
                         compute_unit_id = str(compute_unit["_id"])
@@ -144,7 +142,9 @@ class OutputFileTransferWorker(threading.Thread):
                                        'FTW_Output_Directives.target': sd['target'],
                                        },
                                 update={'$set': {'FTW_Output_Directives.$.state': EXECUTING},
-                                        '$push': {'log': 'Starting transfer of %s' % source}
+                                        '$push': {'log': {
+                                            'timestamp': datetime.datetime.utcnow(),
+                                            'logentry': 'Starting transfer of %s' % source}}
                                 }
                             )
 
@@ -156,8 +156,6 @@ class OutputFileTransferWorker(threading.Thread):
                                 abs_target = "file://localhost%s" % os.path.abspath(target)
 
                             log_msg = "Transferring output file %s -> %s" % (abs_source, abs_target)
-                            logmessage = "Transferred output file %s -> %s" % (abs_source, abs_target)
-                            log_messages.append(log_msg)
                             logger.debug(log_msg)
 
                             logger.debug ("saga.fs.File ('%s')" % saga.Url(abs_source))
@@ -182,22 +180,25 @@ class OutputFileTransferWorker(threading.Thread):
                                        'FTW_Output_Directives.target': sd['target'],
                                        },
                                 update={'$set': {'FTW_Output_Directives.$.state': DONE},
-                                        '$push': {'log': logmessage}
+                                        '$push': {'log': {
+                                            'timestamp': datetime.datetime.utcnow(),
+                                            'logentry': log_msg}}
                                 }
                             )
 
                     except Exception, ex:
                         # Update the CU's state to 'FAILED'.
                         ts = datetime.datetime.utcnow()
-                        log_messages = "Output transfer failed: %s\n%s" % (str(ex), traceback.format_exc())
+                        log_message = "Output transfer failed: %s\n%s" % (str(ex), traceback.format_exc())
                         # TODO: not only mark the CU as failed, but also the specific Directive
                         um_col.update(
-                            {"_id": ObjectId(compute_unit_id)},
-                            {"$set": {"state": FAILED},
-                             "$push": {"statehistory": {"state": FAILED, "timestamp": ts}},
-                             "$push": {"log": log_messages}}
+                            {'_id': ObjectId(compute_unit_id)},
+                            {'$set': {'state': FAILED},
+                             '$push': {'statehistory': {'state': FAILED, 'timestamp': ts}},
+                             '$push': {'log': {'logentry': log_message, 'timestamp': ts}}
+                            }
                         )
-                        logger.error(log_messages)
+                        logger.error(log_message)
 
 
                 # Code below is only to be run by the "first" or only worker
@@ -225,7 +226,11 @@ class OutputFileTransferWorker(threading.Thread):
                         # All Output Directives for this FTW are done, mark the CU accordingly
                         um_col.update({"_id": ObjectId(cu["_id"])},
                                       {'$set': {'FTW_Output_Status': DONE},
-                                       '$push': {'log': 'All FTW output staging directives done - %d.' % self._worker_number}})
+                                       '$push': {'log': {
+                                           'timestamp': datetime.datetime.utcnow(),
+                                           'logentry': 'All FTW output staging directives done - %d.' % self._worker_number}}
+                                       }
+                        )
 
                     # See if there are any Agent Output Directives still pending
                     if cu['Agent_Output_Status'] == EXECUTING and \
@@ -233,8 +238,11 @@ class OutputFileTransferWorker(threading.Thread):
                         # All Output Directives for this Agent are done, mark the CU accordingly
                         um_col.update({"_id": ObjectId(cu["_id"])},
                                       {'$set': {'Agent_Output_Status': DONE},
-                                       '$push': {'log': 'All Agent Output Staging Directives done-%d.' % self._worker_number}
-                                      })
+                                       '$push': {'log': {
+                                           'timestamp': datetime.datetime.utcnow(),
+                                           'logentry': 'All Agent Output Staging Directives done-%d.' % self._worker_number}}
+                                      }
+                        )
 
                 #
                 # Check for all CUs if both Agent and FTW staging is done, we can then mark the CU Done
@@ -256,7 +264,7 @@ class OutputFileTransferWorker(threading.Thread):
                 )
 
         except SystemExit as e :
-            print "output file transfer thread caught system exit -- forcing application shutdown"
+            logger.exception("output file transfer thread caught system exit -- forcing application shutdown")
             import thread
             thread.interrupt_main ()