+++ src/radical/pilot/controller/unit_manager_controller.py
@@ -306,7 +306,7 @@ class UnitManagerController(threading.Thread):
 
 
         except SystemExit as e :
-            print "unit manager controller thread caught system exit -- forcing application shutdown"
+            logger.exception ("unit manager controller thread caught system exit -- forcing application shutdown")
             import thread
             thread.interrupt_main ()
 
@@ -576,7 +576,7 @@ class UnitManagerController(threading.Thread):
                         logger.error('Not sure if action %s makes sense for output staging' % action)
 
                 if unit.FTW_Input_Directives or unit.Agent_Input_Directives:
-                    log = ["Scheduled for data transfer to ComputePilot %s." % pilot_uid]
+                    log = "Scheduled for data transfer to ComputePilot %s." % pilot_uid
                     self._db.set_compute_unit_state(unit.uid, PENDING_INPUT_STAGING, log)
                     cu_transfer.append(unit)
                 else:
@@ -596,7 +596,7 @@ class UnitManagerController(threading.Thread):
             )
 
             for unit in cu_notransfer:
-                log = ["Scheduled for execution on ComputePilot %s." % pilot_uid]
+                log = "Scheduled for execution on ComputePilot %s." % pilot_uid
                 self._db.set_compute_unit_state(unit.uid, PENDING_EXECUTION, log)
                 #self._set_state(uid, PENDING_EXECUTION, log)
 
@@ -605,7 +605,21 @@ class UnitManagerController(threading.Thread):
                 (cu_notransfer, pilot_uid)
             )
         except Exception, e:
-            import traceback
-            logger.error (traceback.format_exc())
-            raise Exception('error in unit manager controler: %s' % e)
+            logger.exception ('error in unit manager controller (schedule())')
+            raise
+
+    # ------------------------------------------------------------------------
+    #
+    def unschedule_compute_units(self, units):
+        """
+        set the unit state to UNSCHEDULED
+        """
+
+        try:
+            unit_ids = [unit.uid for unit in units]
+            self._db.set_compute_unit_state(unit_ids, UNSCHEDULED, "unit remains unscheduled")
+
+        except Exception, e:
+            logger.exception ('error in unit manager controller (unschedule())')
+            raise