+++ src/radical/pilot/pilot_manager.py
@@ -78,7 +78,7 @@ class PilotManager(Object):
                   pm = radical.pilot.PilotManager(session=s)
 
                   pd = radical.pilot.ComputePilotDescription()
-                  pd.resource = "futuregrid.INDIA"  # defined in futuregrid.json
+                  pd.resource = "futuregrid.india"  # defined in futuregrid.json
                   pd.cores    = 16
                   pd.runtime  = 5 # minutes
 
@@ -221,7 +221,6 @@ class PilotManager(Object):
 
         obj = cls(session=session, _reconnect=True)
         obj._uid = pilot_manager_id
-        obj._resource_cfgs = None  # TODO: reconnect
 
         # Retrieve or start a worker process fo this PilotManager instance.
         worker = session._process_registry.retrieve(pilot_manager_id)
@@ -302,15 +301,16 @@ class PilotManager(Object):
                 error_msg = "ComputePilotDescription does not define mandatory attribute 'cores'."
                 raise BadParameter(error_msg)
 
-            # Make sure resource key is known.
-            rcs          = self._session.get_resource_configs()
             resource_key = pilot_description.resource
+            resource_cfg = self._session.get_resource_config(resource_key)
 
-            if resource_key not in rcs:
-                error_msg = "ComputePilotDescription.resource key '%s' is not known by this PilotManager." % resource_key
-                raise BadParameter(error_msg)
-            else:
-                resource_cfg = rcs[resource_key]
+            # Check resource-specific mandatory attributes
+            if "mandatory_args" in resource_cfg:
+                for ma in resource_cfg["mandatory_args"]:
+                    if getattr(pilot_description, ma) is None:
+                        error_msg = "ComputePilotDescription does not define attribute '{0}' which is required for '{1}'.".format(ma, resource_key)
+                        raise BadParameter(error_msg)
+
 
             # we expand and exchange keys in the resource config, depending on
             # the selected schema so better use a deep copy...
@@ -333,14 +333,6 @@ class PilotManager(Object):
                     # merge schema specific resource keys into the
                     # resource config
                     resource_cfg[key] = resource_cfg[schema][key]
-
-            warn_on_local = resource_cfg.get (bool('warn_on_local'), False)
-
-            if  schema == 'local' and warn_on_local :
-                logger.error ("===========================================================")
-                logger.error ("you are using a cluster headnode -- is this what you want??")
-                logger.error ("===========================================================")
-                time.sleep (3)
 
             # If 'default_sandbox' is defined, set it.
             if pilot_description.sandbox is not None: