+++ src/radical/pilot/scheduler/backfilling.py
@@ -98,7 +98,7 @@ class BackfillingScheduler(Scheduler):
                 pid = unit.execution_details.get ('pilot', None)
 
                 if  not pid :
-                    raise RuntimeError ('cannot handle final unit %s w/o pilot information' % uid)
+                    logger.warning ('cannot handle final unit %s w/o pilot information' % uid)
 
                 if  pid not in self.pilots :
                     logger.warning ('cannot handle unit %s cb for pilot %s (pilot is gone)' % (uid, pid))
@@ -149,6 +149,8 @@ class BackfillingScheduler(Scheduler):
                 # FIXME: how can I *un*register a pilot callback?
     
         except Exception as e :
+          # import traceback
+          # traceback.print_exc ()
             logger.error ("error in pilot callback for backfiller (%s) - ignored" % e)
 
 
@@ -190,9 +192,6 @@ class BackfillingScheduler(Scheduler):
         if  not pid in self.pilots :
             raise RuntimeError ('cannot remove unknown pilot (%s)' % pid)
 
-        # NOTE: we don't care if that pilot had any CUs active -- its up to the
-        # UM what happens to those.
-
         del self.pilots[pid]
         # FIXME: how can I *un*register a pilot callback?
 
@@ -213,8 +212,8 @@ class BackfillingScheduler(Scheduler):
             if  unit in self.runq :
                 raise RuntimeError ('Unit cannot be scheduled twice (%s)' % unit.uid)
 
-            if  unit.state != NEW :
-                raise RuntimeError ('Unit %s not in NEW state (%s)' % unit.uid)
+            if  unit.state not in [NEW, UNSCHEDULED] :
+                raise RuntimeError ('Unit %s not in NEW or UNSCHEDULED state (%s)' % unit.uid)
 
             self.waitq.append (unit)
 
@@ -303,8 +302,8 @@ class BackfillingScheduler(Scheduler):
                       # logger.debug ("        unit  %s fits on pilot %s" % (uid, pid))
 
                         # sanity check on unit state
-                        if  unit.state not in [NEW] :
-                            raise RuntimeError ("scheduler queue should only contain NEW units (%s)" % uid)
+                        if  unit.state not in [NEW, UNSCHEDULED] :
+                            raise RuntimeError ("scheduler queue should only contain NEW or UNSCHEDULED units (%s)" % uid)
 
                         self.pilots[pid]['caps'] -= ud.cores
                         schedule['units'][unit]   = pid