+++ src/radical/pilot/tests/api/test_pilot_manager.py
@@ -99,7 +99,7 @@ class Test_PilotManager(unittest.TestCase):
 
         for i in range(0, 2):
             cpd = radical.pilot.ComputePilotDescription()
-            cpd.resource = "localhost"
+            cpd.resource = "local.localhost"
             cpd.cores = 1
             cpd.runtime = 1
             cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"
@@ -128,7 +128,7 @@ class Test_PilotManager(unittest.TestCase):
 
         for i in range(0, 2):
             cpd = radical.pilot.ComputePilotDescription()
-            cpd.resource = "localhost"
+            cpd.resource = "local.localhost"
             cpd.cores = 1
             cpd.runtime = 1
             cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"
@@ -164,7 +164,7 @@ class Test_PilotManager(unittest.TestCase):
 
         for i in range(0, 2):
             cpd = radical.pilot.ComputePilotDescription()
-            cpd.resource = "localhost"
+            cpd.resource = "local.localhost"
             cpd.cores = 1
             cpd.runtime = 1
             cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"
@@ -201,14 +201,14 @@ class Test_PilotManager(unittest.TestCase):
         pmgr = radical.pilot.PilotManager(session=session)
         
         cpd1 = radical.pilot.ComputePilotDescription()
-        cpd1.resource = "localhost"
+        cpd1.resource = "local.localhost"
         cpd1.cores = 1
         cpd1.runtime = 1
         cpd1.sandbox = "/tmp/radical.pilot.sandbox.unittests"
         cpd1.cleanup = True
 
         cpd2 = radical.pilot.ComputePilotDescription()
-        cpd2.resource = "localhost"
+        cpd2.resource = "local.localhost"
         cpd2.cores = 1
         cpd2.runtime = 2
         cpd2.sandbox = "/tmp/radical.pilot.sandbox.unittests"