+++ src/radical/pilot/tests/api/test_unit.py
@@ -57,7 +57,7 @@ class TestUnit(unittest.TestCase):
         pm = radical.pilot.PilotManager(session=session)
 
         cpd = radical.pilot.ComputePilotDescription()
-        cpd.resource = "localhost"
+        cpd.resource = "local.localhost"
         cpd.cores = 1
         cpd.runtime = 1
         cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"
@@ -102,7 +102,7 @@ class TestUnit(unittest.TestCase):
         pm = radical.pilot.PilotManager(session=session)
 
         cpd = radical.pilot.ComputePilotDescription()
-        cpd.resource = "localhost"
+        cpd.resource = "local.localhost"
         cpd.cores = 1
         cpd.runtime = 60
         cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"
@@ -153,7 +153,7 @@ class TestUnit(unittest.TestCase):
         pm = radical.pilot.PilotManager(session=session)
 
         cpd = radical.pilot.ComputePilotDescription()
-        cpd.resource = "localhost"
+        cpd.resource = "local.localhost"
         cpd.cores = 1
         cpd.runtime = 60
         cpd.sandbox = "/tmp/radical.pilot.sandbox.unittests"