+++ src/radical/pilot/utils/db_utils.py
@@ -1,10 +1,43 @@
 
+import os
+import time
 import datetime
 import pymongo
+
+import radical.utils as ru
+
+
+_CACHE_BASEDIR = '/tmp/rp_cache/'
+
+
+# ------------------------------------------------------------------------------
+#
+def bson2json (bson_data) :
+
+    # thanks to
+    # http://stackoverflow.com/questions/16586180/typeerror-objectid-is-not-json-serializable
+
+    import json
+    from   bson.objectid import ObjectId
+
+    class JSONEncoder (json.JSONEncoder) :
+        def default (self, o):
+            if  isinstance (o, ObjectId) :
+                return str (o)
+            if  isinstance (o, datetime.datetime) :
+                seconds  = time.mktime (o.timetuple ())
+                seconds += (o.microsecond / 1000000.0) 
+                return seconds
+            return json.JSONEncoder.default (self, o)
+
+    return ru.parse_json (JSONEncoder ().encode (bson_data))
+
 
 # ------------------------------------------------------------------------------
 #
 def get_session_ids (db) :
+
+    # this is not bein cashed, as the session list can and will change freqently
 
     cnames = db.collection_names ()
     sids   = list()
@@ -24,40 +57,65 @@ def get_last_session (db) :
 
 
 # ------------------------------------------------------------------------------
-def get_session_docs (db, session) :
+def get_session_docs (db, sid, cache=None) :
 
-    ret = dict()
+    # session docs may have been cached in /tmp/rp_cache/<sid>.json -- in that
+    # case we pull it from there instead of the database, which will be much
+    # quicker.  Also, we do cache any retrieved docs to that place, for later
+    # use.
+    if  not cache :
+        cache = "%s/%s.json" % (_CACHE_BASEDIR, sid)
+    else :
+        if  not os.path.isfile (cache) :
+            print "cache '%s' does not exist" % cache
+            return None
 
-    ret['session'] = list(db["%s"    % session].find ())
-    ret['pmgr'   ] = list(db["%s.pm" % session].find ())
-    ret['pilot'  ] = list(db["%s.p"  % session].find ())
-    ret['umgr'   ] = list(db["%s.um" % session].find ())
-    ret['unit'   ] = list(db["%s.cu"  % session].find ())
+    if  os.path.isfile (cache) :
+        return ru.read_json (cache)
 
-    if  len(ret['session']) == 0 :
-        raise ValueError ('no such session %s' % session)
 
-  # if  len(ret['session']) > 1 :
-  #     print 'more than one session document -- pick first one'
+    # cache not used or not found -- go to db
+    bson_data = dict()
+
+    bson_data['session'] = list(db["%s"    % sid].find ())
+    bson_data['pmgr'   ] = list(db["%s.pm" % sid].find ())
+    bson_data['pilot'  ] = list(db["%s.p"  % sid].find ())
+    bson_data['umgr'   ] = list(db["%s.um" % sid].find ())
+    bson_data['unit'   ] = list(db["%s.cu" % sid].find ())
+
+    if  len(bson_data['session']) == 0 :
+        raise ValueError ('no such session %s' % sid)
+
+  # if  len(bson_data['session']) > 1 :
+  #     print 'more than one session document -- picking first one'
 
     # there can only be one session, not a list of one
-    ret['session'] = ret['session'][0]
+    bson_data['session'] = bson_data['session'][0]
 
     # we want to add a list of handled units to each pilot doc
-    for pilot in ret['pilot'] :
+    for pilot in bson_data['pilot'] :
 
         pilot['unit_ids'] = list()
 
-        for unit in ret['unit'] :
+        for unit in bson_data['unit'] :
 
             if  unit['pilot'] == str(pilot['_id']) :
                 pilot['unit_ids'].append (str(unit['_id']))
 
-    return ret
+    # convert bson to json, i.e. serialize the ObjectIDs into strings.
+    json_data = bson2json (bson_data)
+
+
+    # if we got here, we did not find a cached version -- thus add this dataset
+    # to the cache
+    os.system ('mkdir -p %s' % _CACHE_BASEDIR)
+    ru.write_json (json_data, "%s/%s.json" % (_CACHE_BASEDIR, sid))
+
+    return json_data
 
 
 # ------------------------------------------------------------------------------
-def get_session_slothist (db, session) :
+def get_session_slothist (db, sid, cache=None) :
     """
     For all pilots in the session, get the slot lists and slot histories. and
     return as list of tuples like:
@@ -66,7 +124,7 @@ def get_session_slothist (db, session) :
       tuple (string  , list (tuple (string  , int    ) ), list (tuple (string   , datetime ) ) )
     """
 
-    docs = get_session_docs (db, session)
+    docs = get_session_docs (db, sid, cache)
 
     ret = list()
 
@@ -94,7 +152,7 @@ def get_session_slothist (db, session) :
 
 
 # ------------------------------------------------------------------------------
-def get_session_events (db, session) :
+def get_session_events (db, sid, cache=None) :
     """
     For all entities in the session, create simple event tuples, and return
     them as a list
@@ -104,12 +162,12 @@ def get_session_events (db, session) :
       
     """
 
-    docs = get_session_docs (db, session)
+    docs = get_session_docs (db, sid, cache)
 
     ret = list()
 
     if  'session' in docs :
-        doc = docs['session']
+        doc   = docs['session']
         odoc  = dict()
         otype = 'session'
         oid   = str(doc['_id'])